+++ frontend/src/app/pages/home/components/ventanaemerg/ventanaemerg.component.html
@@ -13,7 +13,7 @@
     <!--<h3>Subir Archivos</h3>-->
     <input type="file" (change)="fileProgress($event)" #fileInput style="margin: 1em 0">
     <!--(click)="fileInput.click()"-->
-    <button type="button" class="btn btn-success" style="margin-bottom: 2%;" (click)="openModal('fileDescription')">        
+    <button type="button" class="btn btn-success" style="margin-bottom: 2%;" (click)="onUpload('')">        
         <span class="glyphicon glyphicon-upload"></span> Sube tus archivos
     </button>
     <button type="button" class="btn btn-danger" (click)="clearFiles()" style="margin-left: 2%; margin-bottom: 2%;">
@@ -22,7 +22,7 @@
 
     <jw-modal id="fileDescription">
         <!--<p>Nombre del fichero: <span id="fileName"></span></p>-->
-        <p>Descripción: <input type="text" /></p>
+        <p>Descripción: <input type="text" id="fileDescriptionContent"/></p>
         <button (click)="closeModal('fileDescription')">Aceptar</button>
     </jw-modal>
