+++ frontend/src/app/pages/home/home.component.ts
@@ -157,63 +157,4 @@ export class HomeComponent implements OnInit {
   hideDownloadButton(buttonId: any) {
     document.getElementById('downloadButton' + buttonId).style.display = 'none';
   }
-
-fileData: File = null;
-previewUrl:any = null;
-fileUploadProgress: string = null;
-uploadedFilePath: string = null;
- 
-fileProgress(fileInput: any) {
-      this.fileData = <File>fileInput.target.files[0];
-      this.preview();
-}
- 
-preview() {
-    // Show preview 
-    var mimeType = this.fileData.type;
-    if (mimeType.match(/image\/*/) == null) {
-      return;
-    }
- 
-    var reader = new FileReader();      
-    reader.readAsDataURL(this.fileData); 
-    reader.onload = (_event) => { 
-      this.previewUrl = reader.result; 
-    }
-}
- 
-
-postFile(fileToUpload: File, clientId: any, description: any): Observable<HttpEvent<any>> {
-
-  console.log("postFile");
-
-  const endpoint = `http://localhost:3000/api/Clients/${clientId}/uploadDocument`;
-  const formData: FormData = new FormData();
-  formData.append('file', fileToUpload, fileToUpload.name);
-  formData.append('description', description);
-
-
-  let params = new HttpParams();
-  let headers = new HttpHeaders();
-
-  headers.append("Content-Type", "application/json");
-  headers.append("Content-Type", "multipart/form-data");
-
-  const options = {
-    params: params,
-    reportProgress: true,
-    headers: headers
-  };
-
-  const req = new HttpRequest('POST', endpoint, formData, options);
-
-  return this.http.request(req);
-  }
-
-onUpload() {
-  console.log("onUpload");
-  this.postFile(this.fileData, localStorage.getItem('currentUser'), "descripcion aqui ").subscribe((document) => {
-    /* AQUI YA SE HA SUBIDO EL FICHERO. RECARGAR LISTA Y DEMASES. */
-  });
-}
 }
